refactor : 이미지 패키지 리팩토링
