+++ backend/src/main/java/com/backend/domain/image/domain/repository/ImageJpaRepository.java
@@ -1,0 +1,10 @@
+package com.backend.domain.image.domain.repository;
+
+import com.backend.domain.image.domain.entity.Image;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ImageJpaRepository extends JpaRepository<Image, Long> {
+    public List<Image> findAllByLocationId(Long location_id);
+}
