+++ backend/src/main/java/com/backend/domain/image/dto/ImageResponse.java
@@ -1,0 +1,19 @@
+package com.backend.domain.image.dto;
+
+import com.backend.domain.image.domain.entity.Image;
+import lombok.Builder;
+
+@Builder
+public class ImageResponse {
+
+    private Long id;
+
+    private String path;
+
+    public static ImageResponse of(Image image) {
+        return ImageResponse.builder()
+                .id(image.getId())
+                .path(image.getPath())
+                .build();
+    }
+}
