+++ backend/src/main/java/com/backend/domain/image/exception/ImageNotFoundException.java
@@ -1,0 +1,5 @@
+package com.backend.domain.image.exception;
+
+public class ImageNotFoundException extends RuntimeException{
+
+}
