+++ backend/src/main/java/com/backend/domain/location/domain/entity/Region.java
@@ -1,0 +1,4 @@
+package com.backend.domain.location.domain.entity;
+
+public enum Region {
+}
