+++ backend/src/main/java/com/backend/domain/location/domain/repository/JpaLocationRepository.java
@@ -1,0 +1,10 @@
+package com.backend.domain.location.domain.repository;
+
+import com.backend.domain.location.domain.entity.Location;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface JpaLocationRepository extends JpaRepository<Location, Long> {
+    List<Location> findByTripId(Long tripId);
+}
