+++ backend/src/main/java/com/backend/domain/location/dto/LocationResponse.java
@@ -1,0 +1,4 @@
+package com.backend.domain.location.dto;
+
+public class LocationResponse {
+}
