+++ backend/src/main/java/com/backend/domain/user/controller/UserController.java
@@ -1,0 +1,24 @@
+package com.backend.domain.user.controller;
+
+import com.backend.domain.user.service.UserService;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+
+@Controller // This means that this class is a Controller
+@RequestMapping(path = "/users") // This means URL's start with /demo (after Application path)
+public class UserController {
+    private final UserService userService;
+
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @GetMapping("/search-all")
+    public void search() {
+
+    }
+
+
+}
