+++ backend/src/main/java/com/backend/domain/user/domain/entity/User.java
@@ -1,0 +1,26 @@
+package com.backend.domain.user.domain.entity;
+
+import com.backend.global.BaseTimeEntity;
+import lombok.*;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Getter
+@Setter
+@RequiredArgsConstructor
+@Entity // This tells Hibernate to make a table out of this class
+public class User extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    private String email;
+
+    private String pw;
+
+}
