--- backend/src/main/java/com/backend/entity/BaseTimeEntity.java
@@ -1,24 +1,0 @@
-package com.backend.entity;
-
-import java.time.LocalDateTime;
-
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import lombok.Getter;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public class BaseTimeEntity {
-	@CreatedDate
-	private LocalDateTime createdDate;
-
-	@LastModifiedDate
-	private LocalDateTime modifiedDate;
-
-}
