--- backend/src/main/java/com/backend/entity/Region.java
@@ -1,4 +1,0 @@
-package com.backend.entity;
-
-public enum Region {
-}
