--- backend/src/main/java/com/backend/entity/User.java
@@ -1,25 +1,0 @@
-package com.backend.entity;
-
-import lombok.*;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Getter
-@Setter
-@RequiredArgsConstructor
-@Entity // This tells Hibernate to make a table out of this class
-public class User extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    private String email;
-
-    private String pw;
-
-}
