+++ backend/src/main/java/com/backend/infra/aws/component/ImageUploader.java
@@ -1,0 +1,16 @@
+package com.backend.infra.aws.component;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.Optional;
+
+public interface ImageUploader {
+
+    String upload(MultipartFile multipartFile) throws IOException;
+
+    String changeFileName(MultipartFile uploadFile);
+
+    Optional<File> localUpload(MultipartFile file) throws IOException;
+}
